[diagnostics] Align timing print format strings with compiled binary
Sync format specifiers across all instrumented passes to match the
binary exactly:

- %ld -> %lld for all millisecond values (drop (long) cast; chrono
  milliseconds::count() returns long long)
- %d/%ld -> %u/%lld for Conv2d dims and spatial in cfg-timing
- %-40s -> %-36s for op name field in mla-timing, prop-timing,
  valfd-timing
- runPipeline done uses single format with %s arg for OK/FAIL
- lta-timing and mla-timing shape printing changed to incremental
  fprintf per dimension (matching binary approach)

+++ lib/Dialect/TTNN/Transforms/OptimizerPasses/GreedyL1SpillManagement.cpp
@@ -79,10 +79,10 @@ public:
           func, deviceGrid, l1BudgetPerCore, chipDesc.getUsableL1Size(),
           std::move(observer));
       spill.run();
-      fprintf(stderr, "[l1spill-timing]   func[%zu] '%s' spill.run() done  %ld ms\n",
+      fprintf(stderr, "[l1spill-timing]   func[%zu] '%s' spill.run() done  %lld ms\n",
               funcIdx,
               func.getName().str().c_str(),
-              (long)std::chrono::duration_cast<std::chrono::milliseconds>(
+              std::chrono::duration_cast<std::chrono::milliseconds>(
                   std::chrono::steady_clock::now() - _tSpill)
                   .count());
       ++funcIdx;
@@ -121,8 +121,8 @@ public:
 
       return WalkResult::advance();
     });
-    fprintf(stderr, "[l1spill-timing] GreedyL1SpillManagement TOTAL  %ld ms\n",
-            (long)std::chrono::duration_cast<std::chrono::milliseconds>(
+    fprintf(stderr, "[l1spill-timing] GreedyL1SpillManagement TOTAL  %lld ms\n",
+            std::chrono::duration_cast<std::chrono::milliseconds>(
                 std::chrono::steady_clock::now() - _tTotal)
                 .count());
 #endif
